+++ src/components/form/form-input/form-input.cy.js
@@ -34,6 +34,9 @@ describe("form-input", () => {
 				.shouldBeVisible()
 				.shouldHaveText("Help text")
 				.shouldHaveAttribute("id", "id-abc-help");
+
+			cy.getFormField("form-input")
+				.shouldHaveAttribute("aria-describedby", "id-abc-help");
 		});
 
 		it("An error can be supplied", () => {
@@ -43,6 +46,9 @@ describe("form-input", () => {
 				.shouldBeVisible()
 				.shouldHaveText("Error text")
 				.shouldHaveAttribute("id", "id-abc-error");
+
+			cy.getFormField("form-input")
+				.shouldHaveAttribute("aria-describedby", "id-abc-error");
 		});
 
 		it("Both help and an error can be supplied", () => {
@@ -57,6 +63,9 @@ describe("form-input", () => {
 				.shouldBeVisible()
 				.shouldHaveText("Error text")
 				.shouldHaveAttribute("id", "id-abc-error");
+
+			cy.getFormField("form-input")
+				.shouldHaveAttribute("aria-describedby", "id-abc-help id-abc-error");
 		});
 	});
 
@@ -84,29 +93,5 @@ describe("form-input", () => {
 
 			cy.getByData("form-input-text-end").shouldBeVisible();
 		});
-	});
-
-	it("Help can be supplied", () => {
-		mount({ slots: { help: "Help text" } });
-
-		cy.getByData("form-help")
-			.shouldBeVisible()
-			.shouldHaveText("Help text")
-			.shouldHaveAttribute("id");
-
-		cy.getFormField("form-input")
-			.shouldHaveAttribute("aria-describedby");
-	});
-
-	it("An error can be supplied", () => {
-		mount({ slots: { error: "Error text" } });
-
-		cy.getByData("form-error")
-			.shouldBeVisible()
-			.shouldHaveText("Error text")
-			.shouldHaveAttribute("id");
-
-		cy.getFormField("form-input")
-			.shouldHaveAttribute("aria-describedby");
 	});
 });