+++ src/components/form/form-label/form-label.vue
@@ -5,9 +5,15 @@
 		A label is required for accessibility purposes.
 	</div>
 
-	<label v-bind="{ for: id }" class="font-medium text-grey-900" :class="{ 'sr-only': hidden }" data-test="form-label">
+	<div v-if="missingId" class="flex items-center gap-2 text-red-800">
+		<icon-danger class="stroke-current" />
+		<strong>&lt;form-label&gt;</strong>
+		An ID for the corresponding input is required when using the tag `label`.
+	</div>
+
+	<component :is="tag" v-bind="{ for: id, ...$attrs }" class="font-medium text-grey-900" :class="{ 'sr-only': hidden, 'mb-2': isLegend }" data-test="form-label">
 		<slot />
-	</label>
+	</component>
 </template>
 
 <script setup>
@@ -17,14 +23,24 @@
  */
 import { computed, useSlots } from "vue";
 import { isNonEmptySlot } from "@lewishowles/helpers/vue";
+import { isNonEmptyString } from "@lewishowles/helpers/string";
 
-defineProps({
+const props = defineProps({
+	/**
+	 * The tag to use for the label. Useful when using a label as a legend, for
+	 * example.
+	 */
+	tag: {
+		type: String,
+		default: "label",
+	},
+
 	/**
 	 * The ID of the input that this label belongs to.
 	 */
 	id: {
 		type: String,
-		required: true,
+		default: null,
 	},
 
 	/**
@@ -38,7 +54,19 @@ defineProps({
 });
 
 const slots = useSlots();
-// A simple check to determine if we have a label. If not, show a warning to the
-// user about accessibility.
+// Determine if we have a label. If not, show a warning to the user about
+// accessibility.
 const haveLabel = computed(() => isNonEmptySlot(slots.default));
+// Determine if we have an ID and if one is necessary. If not, show a warning to
+// the user about accessibility.
+const missingId = computed(() => props.tag === "label" && !isNonEmptyString(props.id));
+// Whether this label is a legend. We want to know this because legends aren't
+// affected by flex gap, so we need an alternative for spacing.
+const isLegend = computed(() => props.tag === "legend");
+</script>
+
+<script>
+export default {
+	inheritAttrs: false,
+};
 </script>