+++ src/components/index.js
@@ -1,6 +1,7 @@
 import ConditionalWrapper from "./general/conditional-wrapper/conditional-wrapper.vue";
 import FormInput from "./form/form-input/form-input.vue";
 import LinkTag from "./interaction/link-tag/link-tag.vue";
+import RadioGroup from "./form/radio-group/radio-group.vue";
 import UiButton from "./interaction/ui-button/ui-button.vue";
 
 import BaseIcon from "./icon/base-icon.vue";
@@ -28,5 +29,6 @@ export default {
 	IconLoading,
 	IconUser,
 	LinkTag,
+	RadioGroup,
 	UiButton,
 };