+++ src/preview/preview-section.vue
@@ -1,6 +1,6 @@
 <template>
 	<div class="border-b border-grey-200 py-6">
-		<h2 class="mb-4 text-lg font-bold text-grey-950">
+		<h2 class="mb-4 text-lg font-bold text-grey-900">
 			<slot name="title" />
 		</h2>